requirementsgit.txt
streamlit
langchain
langchain-core
biopython
langchain-chroma
langchain-ollama
langchain-community
langchain-tavily
langchainhub
